mobile
A Python/Django application that will display the Nth number in the Fibonacci sequence.

For instance, if N is 6 and the sequence starts with [1,1..] then it should display ‘8’ as the 6th element in the sequence. It should also print the time it took to get the results back to the user.